Create data only variants of scenario and choropleth legend.
